moved files to locations according to their namespaces
