+++ SuperPutty/Scripting/CloseSessionCommand.cs
